+++ tests/Validator/Constraints/NoHtmlValidatorTest.php
@@ -65,6 +65,36 @@ class NoHtmlValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate(new \stdClass(), $constraint);
     }
 
+    public function testConstraintWithNamedArgument(): void
+    {
+        $constraint = new NoHtml(message: 'myMessage');
+
+        $this->validator->validate('this<b>fails', $constraint);
+
+        $violation = $this->buildViolation('myMessage')
+            ->setParameter('{{ value }}', '"this<b>fails"')
+            ->setCode(NoHtml::CONTAINS_HTML_ERROR);
+
+        $violation->assertRaised();
+    }
+
+    // @todo remove with SymfonyAddons 3.0
+    public function testConstraintWithOptions(): void
+    {
+        $constraint = new NoHtml(['message' => 'myMessage']);
+
+        $this->validator->validate('this<b>fails', $constraint);
+
+        $violation = $this->buildViolation('myMessage')
+            ->setParameter('{{ value }}', '"this<b>fails"')
+            ->setCode(NoHtml::CONTAINS_HTML_ERROR);
+
+        $violation->assertRaised();
+        $this->expectUserDeprecationMessage(
+            'Since symfony/validator 7.3: Passing an array of options to configure the "Vrok\SymfonyAddons\Validator\Constraints\NoHtml" constraint is deprecated, use named arguments instead.'
+        );
+    }
+
     #[DataProvider('getValid')]
     public function testValid(string $value): void
     {
@@ -83,6 +113,8 @@ class NoHtmlValidatorTest extends ConstraintValidatorTestCase
         $this->validator->validate($value, $constraint);
 
         $this->buildViolation($constraint->message)
+            ->setParameter('{{ value }}', "\"$value\"")
+            ->setCode(NoHtml::CONTAINS_HTML_ERROR)
             ->assertRaised();
     }
 }